+++ src/components/MainSection/MovieInfo.js
@@ -1,32 +1,17 @@
 import { memo } from 'react';
 import { useState, useEffect } from 'react';
 import ProviderIconsList from './ProviderIconsList.js';
+import { ProviderIconsApiCall } from '../MovieApi.js';
 import { EyeIcon } from '../Icons.js';
 
 function MovieInfo({ overview, movieTitle, movieID, tvMovieToggle, currentRegion, infoState, setInfoState }) {
 
-    const [viewingOptions, setviewingOptions] = useState([]);
+    const [viewingOptions, setViewingOptions] = useState({});
 
     useEffect(() => {
-        // there is no way to filter by region (https://www.themoviedb.org/talk/643dbcf75f4b7304e2fe7f2a)
-        const getMovieviewingOptions = `https://api.themoviedb.org/3/${tvMovieToggle}/${movieID}/watch/providers?api_key=0a093f521e98a991f4e4cc2a12460255`;
-
-        fetch(getMovieviewingOptions)
-            .then(results => {
-                return results.json();
-            })
-            .then(data => {
-
-                const emptyObject = {
-                    buy_rent: [{ logo_path: 'N/A', provider_id: 'buy/rent_N/A', provider_name: 'N/A' }],
-                    stream: [{ logo_path: 'N/A', provider_id: 'stream_N/A', provider_name: 'N/A' }]
-                }
-
-                setviewingOptions(data.results[currentRegion[0]] ? data.results[currentRegion[0]] : emptyObject);
-            }).catch(() => {
-                console.log("Failed to fetch streaming options");
-            })
-    }, [setviewingOptions, tvMovieToggle, movieID, currentRegion]);
+        ProviderIconsApiCall(tvMovieToggle, movieID, currentRegion).then(result => setViewingOptions(result));
+        console.log(viewingOptions)
+    }, [setViewingOptions, tvMovieToggle, movieID, currentRegion]);
 
     const handleMovieInfo = () => {
         setInfoState('provider-info');
@@ -46,14 +31,8 @@ function MovieInfo({ overview, movieTitle, movieID, tvMovieToggle, currentRegion
                     <p>{overview}</p>
                 </div>
                 <div className={infoState === 'provider-info' ? 'movie-info' : 'hidden'}>
-                    {(!Object.keys(viewingOptions).length > 0) ?
-                        <div className='no-options'>
-                            <h3>No viewing options for</h3>
-                            <h3 className='no-options-movie-title'>{movieTitle}</h3>
-                            <h3>in your region</h3>
-                        </div>
-
-                        : <ProviderIconsList
+                    {
+                        <ProviderIconsList
                             movieTitle={movieTitle}
                             movieID={movieID}
                             viewingOptions={viewingOptions}
